LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/at-position-try-invalidation-shadow-dom-expected.txt


PASS #host is initially left: 999999px
PASS #slotted is initially left: 999999px
PASS #host with inserted @position-try applied
PASS #slotted with inserted @position-try applied
